Implemented loading of DB mapping files from plugins

+++ ultimategdbot-core/src/main/java/com/github/alex1304/ultimategdbot/core/handler/ReplyMenuHandler.java
@@ -16,6 +16,7 @@ import discord4j.core.event.domain.message.MessageCreateEvent;
 import discord4j.core.object.entity.Message;
 import reactor.core.Disposable;
 import reactor.core.publisher.Mono;
+import reactor.core.scheduler.Schedulers;
 
 public class ReplyMenuHandler implements Handler {
 	
@@ -70,7 +71,7 @@ public class ReplyMenuHandler implements Handler {
 
 	@Override
 	public void listen() {
-		bot.getDiscordClient().getEventDispatcher().on(MessageCreateEvent.class)
+		bot.getDiscordClient().getEventDispatcher().on(MessageCreateEvent.class).subscribeOn(Schedulers.elastic())
 				.filter(event -> event.getMessage().getContent().isPresent()
 						&& event.getMessage().getAuthorId().isPresent()
 						&& openedReplyMenus.containsKey(event.getMessage().getChannelId().asString()
@@ -86,7 +87,9 @@ public class ReplyMenuHandler implements Handler {
 					}
 					Command.invoke(action, ctx);
 					replyMenu.complete();
-					ctx.getEvent().getMessage().delete().doOnError(__ -> {}).subscribe();
+					if (replyMenu.deleteOnReply) {
+						ctx.getEvent().getMessage().delete().doOnError(__ -> {}).subscribe();
+					}
 				});
 	}
